outsource, refactor cache2 unit tests
